fix(refactor): extract lowering for issue #226
Move lower_program, lower_node, lower_pattern and related private helpers
from interpreter.py into src/genia/lowering.py. Re-export public names
and _lambda_pattern_is_simple_parameter_shape from interpreter.py for
backward compatibility with existing tests and tools.
